backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
python-dotenv
pygithub
pydantic
requests
